pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "voice-agent-web"
version = "0.1.0"
description = "Browser-based real-time voice agent with local models"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "pywhispercpp>=1.2",
    "piper-tts>=1.4",
    "coqui-tts[codec]>=0.24",
    "transformers>=4.40,<5",
    "torch>=2.2",
    "httpx>=0.28",
    "numpy>=1.26",
    "torchaudio>=2.10.0",
]

[tool.hatch.build.targets.wheel]
packages = ["server"]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.25",
]
